+++ src/post_modules/Shuffler/e_source_block_pikestream.c
@@ -32,7 +32,7 @@ struct pf_source
 };
 
 
-static struct data get_data( struct source *_s, int len )
+static struct data get_data( struct source *_s, off_t len )
 {
   struct pf_source *s = (struct pf_source *)_s;
   struct data res;
